Clean-Up and Finish Calculator Logic
Removed files that are automatically generated when making a Blazor Server project which generate template websites. Left the ones in the WebServiceApp section as we will need them in this section. Finished the calculator logic and minimum unit tests needed according to the requirements documentation on Jeff's GitHub Repo for SWE 3643 Course.