Updated Calculator Webpage
The Calculator.Razor file has all the logic working with no UI. The Home.Razor file has all the UI, but no logic connected. Next Step: Merging the two together.

KEY NOTE: The Calculator.Razor file does not currently handle when exceptions occur. This needs to be fixed for full functionality. As of now, if an error occurs, the page freezes completely.