Included Logic Errors onto Web App.
The Web App portion of our project now handles when errors occur. Modifications were made to the calculator engine and unit tests to include this change. No "errors" are thrown now. They are instead not calculated to ensure the application doesnt freeze while the user is using it!

Weird Errors:
- For some reason the input fields on the web app do not work when the duplicate ones are commented out. Will ask in class about this issue!

Whats Left:
- Playwright Tests (Struggling to install playwright on my Macbook. Will try to figure it out later. If not, I will work on the playwright tests on my home desktop as it is already installed.
-Video Presentation and Github Repo documentation.